[Update] WeatherViewModel 를 통한 지역 날씨 정보 가져오기
지역 날씨 정보는 다음과 같이 사용자에게 보여줍니다.
1. WeatherFragment 에서 WeatherViewModel 를 바인딩 합니다.
2. WeatherViewModel 에 있는 locationWeatherInfoListLiveData 를
WeatherFragment 에서 observe 합니다.
3. WeatherViewModel 에 getLocation() 메서드를 호출합니다.
4. 이후 GetLocationUseCase 를 통해 "se" 가 들어가 있는 지역 리스트를 가져 옵니다.
5. 지역 리스트를 순회하며 각 지역의 id 값을 통해 GetWeatherUseCase 으로
지역 날씨 정보 리스트를 가져옵니다.
6. 지역 날씨 정보 리스트를 locationWeatherInfoListLiveData 의 value 값에 넣습니다.
7. 이후, WeatherFragment 에서 locationWeatherInfoListLiveData 의 observe 에
정의된 작업에 따라 사용자에게 지역 날씨 정보 리스트를 보여줍니다.

변경사항
1. id 값만 있으면 GetWeatherUseCase 에서 지역 이름이랑 날씨 정보를
모두 가져 올 수 있기때문에 LocationEntity 를 삭제하고 id 값과
날씨 정보 리스트를 담은 LocationWeatherInfoEntity 를 추가하였습니다.
때문에 LocationEntity 와 관련된 모든 파일들이 LocationWeatherInfoEntity 에 맞게
수정되었습니다.

+++ data/src/main/java/com/meta/weather/data/entity/WeatherData.kt
@@ -3,6 +3,6 @@ package com.meta.weather.data.entity
 data class WeatherData(
     val weather_state_name: String,
     val weather_state_abbr: String,
-    val the_temp: Float,
+    val the_temp: Double,
     val humidity: Int
 )
